chore: use extracted java plugin
